Code & tests movement.
