+++ abi/codecForCompositeTypes_list.go
@@ -5,34 +5,6 @@ import (
 	"io"
 )
 
-func (c *codec) encodeNestedOption(writer io.Writer, value OptionValue) error {
-	if value.Value == nil {
-		_, err := writer.Write([]byte{0})
-		return err
-	}
-
-	_, err := writer.Write([]byte{1})
-	if err != nil {
-		return err
-	}
-
-	return c.doEncodeNested(writer, value.Value)
-}
-
-func (c *codec) decodeNestedOption(reader io.Reader, value *OptionValue) error {
-	bytes, err := readBytesExactly(reader, 1)
-	if err != nil {
-		return err
-	}
-
-	if bytes[0] == 0 {
-		value.Value = nil
-		return nil
-	}
-
-	return c.doDecodeNested(reader, value.Value)
-}
-
 func (c *codec) encodeNestedList(writer io.Writer, value InputListValue) error {
 	err := encodeLength(writer, uint32(len(value.Items)))
 	if err != nil {
