+++ abi/codec_test.go
@@ -32,40 +32,6 @@ func TestCodec_EncodeNested(t *testing.T) {
 		pubKeyLength: 32,
 	})
 
-	doTest := func(t *testing.T, value any, expected string) {
-		encoded, err := codec.EncodeNested(value)
-		require.NoError(t, err)
-		require.Equal(t, expected, hex.EncodeToString(encoded))
-	}
-
-	t.Run("option with value", func(t *testing.T) {
-		fooOption := OptionValue{
-			Value: U16Value{Value: 0x08},
-		}
-
-		doTest(t, fooOption, "010008")
-	})
-
-	t.Run("option without value", func(t *testing.T) {
-		fooOption := OptionValue{
-			Value: nil,
-		}
-
-		doTest(t, fooOption, "00")
-	})
-
-	t.Run("list", func(t *testing.T) {
-		fooList := InputListValue{
-			Items: []any{
-				U16Value{Value: 1},
-				U16Value{Value: 2},
-				U16Value{Value: 3},
-			},
-		}
-
-		doTest(t, fooList, "00000003000100020003")
-	})
-
 	t.Run("should err when unknown type", func(t *testing.T) {
 		type dummy struct {
 			foobar string
@@ -127,52 +93,6 @@ func TestCodec_DecodeNested(t *testing.T) {
 		destination := &BigIntValue{}
 		err := codec.DecodeNested(data, destination)
 		require.ErrorContains(t, err, "cannot decode (nested) *abi.BigIntValue, because of: cannot read exactly 3 bytes")
-	})
-
-	t.Run("option with value", func(t *testing.T) {
-		data, _ := hex.DecodeString("010008")
-
-		destination := &OptionValue{
-			Value: &U16Value{},
-		}
-
-		err := codec.DecodeNested(data, destination)
-		require.NoError(t, err)
-		require.Equal(t, &OptionValue{
-			Value: &U16Value{Value: 8},
-		}, destination)
-	})
-
-	t.Run("option without value", func(t *testing.T) {
-		data, _ := hex.DecodeString("00")
-
-		destination := &OptionValue{
-			Value: &U16Value{},
-		}
-
-		err := codec.DecodeNested(data, destination)
-		require.NoError(t, err)
-		require.Equal(t, &OptionValue{
-			Value: nil,
-		}, destination)
-	})
-
-	t.Run("list", func(t *testing.T) {
-		data, _ := hex.DecodeString("00000003000100020003")
-
-		destination := &OutputListValue{
-			ItemCreator: func() any { return &U16Value{} },
-			Items:       []any{},
-		}
-
-		err := codec.DecodeNested(data, destination)
-		require.NoError(t, err)
-		require.Equal(t,
-			[]any{
-				&U16Value{Value: 1},
-				&U16Value{Value: 2},
-				&U16Value{Value: 3},
-			}, destination.Items)
 	})
 
 	t.Run("should err when unknown type", func(t *testing.T) {
